コンポーネント指向に変更

+++ Library/DxLibHelper/Vector/vector2_int.hpp
@@ -39,7 +39,7 @@ struct Vector2Int
 	/// @return 大きさの二乗
 	[[nodiscard]] float GetSqrMagnitude() const
 	{
-		return x * x + y * y;
+		return static_cast<float>(x * x + y * y);
 	}
 
 	/// @brief 自身を正規化する
